+++ example/android/app/src/main/java/com/example/flutter_file_downloader_example/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.flutter_file_downloader_example;
+package com.odehbros.flutter_file_downloader_example;
 
 import io.flutter.embedding.android.FlutterActivity;