+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 1.1.0+1
+## 1.1.1
 
-Updated the README.md
+Fixed issue to increase plugin rate on pub.dev
+Fixed permission issue in android 11+